ai-infra-agent-hackathon-main/requirements.txt
boto3==1.34.144
botocore==1.34.144
langchain==0.3.0
langchain-core>=0.3.27,<0.4.0
langchain-openai==0.2.14
langchain-anthropic==0.2.4
langchain-groq==0.2.0
langchain-community==0.3.0
langchain-text-splitters>=0.3.0,<0.4.0
langsmith>=0.1.17,<0.2.0
anthropic>=0.40.0
openai>=1.30.0,<2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv==1.0.1
streamlit==1.37.1
pytest==8.3.2
markdown==3.10.2
localstack[full]>=3.0.0
awscli-local>=0.22.0
